pyproject.toml
[project]
name = "nike-ad-crawler"
version = "0.1.0"
description = ""
authors = [
    {name = "tuannd4",email = "tuannd4@smartosc.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "weasyprint (>=65.0,<66.0)",
    "markdown (>=3.7,<4.0)",
    "openai (>=1.70.0,<2.0.0)",
    "uuid (>=1.30,<2.0)",
    "streamlit (>=1.44.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "nike_ad_crawler", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
